// frontend/src/types/filter.enum.ts
export enum FilterNumberCondition {
    GREATER_THAN = 'greater_than',
    LESS_THAN = 'less_than',
}

export enum FilterStringCondition {
    EQUAL = 'equal',
    NOT_EQUAL = 'not_equal',
}